Update command-line-scripts.rst
Most answers on Stack Overflow point to this document, which I read up and down and struggled to implement for hours using the 'scripts' keyword approach before finding one little comment on a lower ranked answer in SO that mentioned the 'scripts' keyword route won't work on Windows. That information should really be in this document. Just trying to save someone else the 3 hours I can't get back.